Refactor test for chooseTeam, add test for store
