+++ lib/chooseTeam.js
@@ -28,6 +28,6 @@ const getTeamChoice = async () => {
   }
 };
 
-exports.getTeamChoice = getTeamChoice;
+exports.promptForTeamChoice = promptForTeamChoice;
 exports.isValidTeamChoice = isValidTeamChoice;
-exports.setTeam = setTeam;
+exports.getTeamChoice = getTeamChoice;
