+++ lib/chooseTeam.test.js
@@ -1,9 +1,8 @@
 const rl = require("./rl");
 const {
   getTeamChoice,
-  getValidTeamChoice,
-  setTeam,
-  storeTeamChoice,
+  isValidTeamChoice,
+  promptForTeamChoice,
 } = require("./chooseTeam");
 
 jest.mock("./rl");
@@ -19,19 +18,19 @@ describe("Choices are stored correctly", () => {
     consoleOutput = [];
   });
 
-  it("getTeamChoice returns user input", async () => {
+  it("promptForTeamChoice returns user input", async () => {
     const expected = "X";
 
     rl.question = (question, cb) => {
       cb(expected);
     };
 
-    const result = await getTeamChoice();
+    const result = await promptForTeamChoice();
 
     expect(result).toEqual(expected);
   });
 
-  it("getTeamChoice returns user input in upperCase", async () => {
+  it("promptForTeamChoice returns user input in upperCase", async () => {
     const expected = "X";
     const input = "x";
 
@@ -39,119 +38,62 @@ describe("Choices are stored correctly", () => {
       cb(input);
     };
 
-    const result = await getTeamChoice();
+    const result = await promptForTeamChoice();
     expect(result).toEqual(expected);
   });
 
-  it("getValidTeamChoice returns true", () => {
+  it("isValidTeamChoice returns true", () => {
     const expected = true;
     const input1 = "X";
     const input2 = "O";
 
-    const result1 = getValidTeamChoice(input1);
-    const result2 = getValidTeamChoice(input2);
+    const result1 = isValidTeamChoice(input1);
+    const result2 = isValidTeamChoice(input2);
     expect(result1).toEqual(expected);
     expect(result2).toEqual(expected);
   });
 
-  it("getValidTeamChoice returns false", () => {
+  it("isValidTeamChoice returns false", () => {
     const expected = false;
     const input = "foo";
 
-    const result = getValidTeamChoice(input);
+    const result = isValidTeamChoice(input);
     expect(result).toEqual(expected);
   });
 
-  it("getValidTeamChoice validates getTeamChoice of 'X'", async () => {
-    const expected = true;
-    const input = "X";
-    rl.question = (question, cb) => {
-      cb(input);
-    };
-    const result = getValidTeamChoice(await getTeamChoice());
-    expect(result).toEqual(expected);
-  });
-
-  it("getValidTeamChoice validates getTeamChoice of 'foo'", async () => {
-    const expected = false;
-    const input = "foo";
-    rl.question = (question, cb) => {
-      cb(input);
-    };
-    const result = getValidTeamChoice(await getTeamChoice());
-    expect(result).toEqual(expected);
-  });
-
-  it("setTeam stores team choice of 'X'", () => {
-    const store = {
-      userTeam: null,
-      cpuTeam: null,
-    };
-    const expected = {
-      userTeam: "X",
-      cpuTeam: "O",
-    };
-    setTeam(store, "X");
-    expect(store).toEqual(expected);
-  });
-
-  it("setTeam stores team choice of 'O'", () => {
-    const store = {
-      userTeam: null,
-      cpuTeam: null,
-    };
-    const expected = {
-      userTeam: "O",
-      cpuTeam: "X",
-    };
-    setTeam(store, "O");
-    expect(store).toEqual(expected);
-  });
-
-  it("storeTeamChoice validates and stores choice of 'X'", async () => {
-    const store = {
-      userTeam: null,
-      cpuTeam: null,
-    };
-    const expected = {
-      userTeam: "X",
-      cpuTeam: "O",
-    };
+  it("getTeamChoice returns a valid team choice", async () => {
+    const expected = "X";
     const input = "X";
 
     rl.question = (question, cb) => {
       cb(input);
     };
 
-    await storeTeamChoice(store);
-    expect(store).toEqual(expected);
+    const result = await getTeamChoice();
+    expect(result).toEqual(expected);
   });
 
-  it("storeTeamChoice validates and rejects choice of 'foo'", async () => {
-    const store = {
-      userTeam: null,
-      cpuTeam: null,
-    };
-    const expected = {
-      userTeam: "X",
-      cpuTeam: "O",
-    };
+  it("getTeamChoice rejects until given a valid team choice", async () => {
+    const expected = "X";
     const expectedLog =
       "Please enter only the letter O or X to choose your team";
-    const input = "foo";
 
-    rl.question = (question, cb) => {
-      cb(input);
-    };
+    rl.question = jest
+      .fn()
+      .mockImplementationOnce((question, cb) => {
+        cb("foo");
+      })
+      .mockImplementationOnce((question, cb) => {
+        cb("X");
+      })
+      .mockImplementationOnce((question, cb) => {
+        cb("bar");
+      });
 
-    const mock = jest.fn();
-    await storeTeamChoice(store);
+    const result = await getTeamChoice();
 
-    expect(mock).toHaveBeenCalled();
-    mock.mockRestore();
-
-    // expect(storeTeamChoice).toHaveBeenCalled();
+    expect(result).toEqual(expected);
+    expect(rl.question).toHaveBeenCalledTimes(2);
     expect(consoleOutput).toEqual([expectedLog]);
-    // expect(store).toEqual(expected);
   });
 });
