+++ lib/store.js
@@ -1,3 +1,5 @@
+const { ModuleNotFoundError } = require("jest-resolve");
+
 const setTeam = (store, choice) => {
   store.userTeam = choice;
   if (choice === "O") {
@@ -6,3 +8,5 @@ const setTeam = (store, choice) => {
     store.cpuTeam = "O";
   }
 };
+
+module.exports = setTeam;
